Move machines to separate files
