+++ examples/readme.md
@@ -2,6 +2,7 @@
 
 1. Copy any of the template directories to this directory:
 
+   - [template-js](./template-js)
    - [template-ts](./template-ts)
    - [template-react-ts](./template-react-ts)
    - [template-vue-ts](./template-vue-ts)
