+++ examples/template-js/src/toggle.machine.js
@@ -1,0 +1,18 @@
+import { createMachine } from 'xstate';
+
+// Edit your machine(s) here
+export const toggleMachine = createMachine({
+  id: 'machine',
+  initial: 'inactive',
+  context: {
+    count: 0
+  },
+  states: {
+    inactive: {
+      on: { TOGGLE: 'active' }
+    },
+    active: {
+      on: { TOGGLE: 'inactive' }
+    }
+  }
+});
