+++ examples/template-ts/src/main.ts
@@ -1,5 +1,6 @@
-import { createMachine, interpret } from 'xstate';
+import { interpret } from 'xstate';
 import './style.css';
+import { machine } from './toggle.machine';
 
 const app = document.querySelector<HTMLDivElement>('#app')!;
 
@@ -8,31 +9,6 @@ app.innerHTML = `
   <p>Open the console!</p>
   <button id="toggle">Toggle</button>
 `;
-
-interface ToggleContext {
-  count: number;
-}
-
-type ToggleEvent = {
-  type: 'TOGGLE';
-};
-
-// Edit your machine(s) here
-const machine = createMachine<ToggleContext, ToggleEvent>({
-  id: 'machine',
-  initial: 'inactive',
-  context: {
-    count: 0
-  },
-  states: {
-    inactive: {
-      on: { TOGGLE: 'active' }
-    },
-    active: {
-      on: { TOGGLE: 'inactive' }
-    }
-  }
-});
 
 const button = document.querySelector<HTMLButtonElement>('#toggle');
