+++ examples/template-vue-ts/src/toggle.machine.ts
@@ -1,0 +1,18 @@
+import { createMachine, assign } from 'xstate';
+
+export const toggleMachine = createMachine<{ count: number }>({
+  id: 'toggle',
+  initial: 'inactive',
+  context: {
+    count: 0
+  },
+  states: {
+    inactive: {
+      on: { TOGGLE: 'active' }
+    },
+    active: {
+      entry: assign({ count: (ctx) => ctx.count + 1 }),
+      on: { TOGGLE: 'inactive' }
+    }
+  }
+});
